requirements.txt
pandas
psycopg2-binary
configparser
tqdm
python-dotenv
pandas_schema